requirements.txt
Django==4.0.4
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1

psycopg2-binary==2.9.3
celery==5.2.6
PyYAML==6.0
redis==4.3.1
requests==2.27.1
django-allauth


